extract diversity_selection form ConditionExampleGenerator
