+++ app/javascript/editorV2/panels/condition_preview/diversity_selection.js
@@ -1,0 +1,114 @@
+import { candidateIdentity } from 'editorV2/panels/condition_preview/example_utils'
+
+export function subjectSpeciesSignature(example) {
+  return example.result.subjectPositions.map(position => example.afterBoard.pieceTypeAt(position)).join(',')
+}
+
+export function targetSpeciesSignature(example) {
+  return example.result.targetPositions.map(position => example.afterBoard.pieceTypeAt(position)).join(',')
+}
+
+export function speciesPairSignature(example) {
+  return `${subjectSpeciesSignature(example)}=>${targetSpeciesSignature(example)}`
+}
+
+export function varietySignature(example) {
+  return [
+    example.variantType,
+    subjectSpeciesSignature(example),
+    targetSpeciesSignature(example),
+    example.geometryKey
+  ].join(':')
+}
+
+export function bucketKeyForExample(example) {
+  return [
+    subjectSpeciesSignature(example),
+    targetSpeciesSignature(example),
+    example.variantType
+  ].join('|')
+}
+
+export function uniqueExamples(examples) {
+  const seen = new Set()
+  return examples.filter(example => {
+    const identity = candidateIdentity(example)
+    if (seen.has(identity)) { return false }
+    seen.add(identity)
+    return true
+  })
+}
+
+export function roundRobinAppend({ selected, candidatesByKey, maxExamples, seenIdentities }) {
+  const queue = Array.from(candidatesByKey.entries()).map(([key, candidates]) => ({
+    key,
+    candidates: [...candidates]
+  }))
+  let added = false
+
+  while (selected.length < maxExamples) {
+    let progressed = false
+
+    for (let index = 0; index < queue.length && selected.length < maxExamples; index += 1) {
+      const bucket = queue[index]
+      while (bucket.candidates.length > 0 && seenIdentities.has(candidateIdentity(bucket.candidates[0]))) {
+        bucket.candidates.shift()
+      }
+      if (bucket.candidates.length === 0) { continue }
+
+      const next = bucket.candidates.shift()
+      selected.push(next)
+      seenIdentities.add(candidateIdentity(next))
+      progressed = true
+      added = true
+    }
+
+    if (!progressed) { break }
+  }
+
+  return added
+}
+
+export function selectDiverseExamples(candidates, maxExamples) {
+  if (candidates.length <= maxExamples) { return [...candidates] }
+
+  const selected = []
+  const seenIdentities = new Set()
+  const subjectBuckets = new Map()
+  const targetBuckets = new Map()
+  const pairBuckets = new Map()
+  const variantBuckets = new Map()
+
+  candidates.forEach(candidate => {
+    const subjectKey = subjectSpeciesSignature(candidate)
+    const targetKey = targetSpeciesSignature(candidate)
+    const pairKey = speciesPairSignature(candidate)
+    const variantKey = candidate.variantType
+
+    if (!subjectBuckets.has(subjectKey)) { subjectBuckets.set(subjectKey, []) }
+    if (!targetBuckets.has(targetKey)) { targetBuckets.set(targetKey, []) }
+    if (!pairBuckets.has(pairKey)) { pairBuckets.set(pairKey, []) }
+    if (!variantBuckets.has(variantKey)) { variantBuckets.set(variantKey, []) }
+
+    subjectBuckets.get(subjectKey).push(candidate)
+    targetBuckets.get(targetKey).push(candidate)
+    pairBuckets.get(pairKey).push(candidate)
+    variantBuckets.get(variantKey).push(candidate)
+  })
+
+  roundRobinAppend({ selected, candidatesByKey: subjectBuckets, maxExamples, seenIdentities })
+  roundRobinAppend({ selected, candidatesByKey: targetBuckets, maxExamples, seenIdentities })
+  roundRobinAppend({ selected, candidatesByKey: pairBuckets, maxExamples, seenIdentities })
+  roundRobinAppend({ selected, candidatesByKey: variantBuckets, maxExamples, seenIdentities })
+
+  if (selected.length >= maxExamples) {
+    return selected.slice(0, maxExamples)
+  }
+
+  const remaining = candidates.filter(candidate => !seenIdentities.has(candidateIdentity(candidate)))
+  for (let index = 0; index < remaining.length && selected.length < maxExamples; index += 1) {
+    selected.push(remaining[index])
+  }
+
+  return selected
+}
